+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Fix: Handle immutable event lists and maps
 - Fix: NDK integration was being disabled by a typo
 - Fix: missing toList for debug meta #192
+- Enhancement: NavigationObserver to record Breadcrumbs for navigation events #197
 
 ### Breaking changes